refactor : 모임상세 페이지 컴포넌트 분리 및 대형화면 스타일링
